+++ core/src/test/java/com/graphhopper/routing/PriorityRoutingTest.java
@@ -74,7 +74,7 @@ public class PriorityRoutingTest {
         // A* and Dijkstra should yield the same path (the max priority must be taken into account by weighting.getMinWeight)
         {
             CustomModel customModel = new CustomModel();
-            CustomWeighting weighting = CustomModelParser.createWeighting(accessEnc,
+            CustomWeighting weighting = CustomModelParser.createLegacyWeighting(accessEnc,
                     speedEnc, priorityEnc, em,
                     TurnCostProvider.NO_TURN_COST_PROVIDER, customModel);
             Path pathDijkstra = new Dijkstra(graph, weighting, TraversalMode.NODE_BASED).calcPath(0, 3);
@@ -87,7 +87,7 @@ public class PriorityRoutingTest {
             CustomModel customModel = new CustomModel();
             // now we even increase the priority in the custom model, which also needs to be accounted for in weighting.getMinWeight
             customModel.addToPriority(Statement.If("road_class == MOTORWAY", Statement.Op.MULTIPLY, "3"));
-            CustomWeighting weighting = CustomModelParser.createWeighting(accessEnc, speedEnc,
+            CustomWeighting weighting = CustomModelParser.createLegacyWeighting(accessEnc, speedEnc,
                     priorityEnc, em, TurnCostProvider.NO_TURN_COST_PROVIDER, customModel);
             Path pathDijkstra = new Dijkstra(graph, weighting, TraversalMode.NODE_BASED).calcPath(0, 3);
             Path pathAStar = new AStar(graph, weighting, TraversalMode.NODE_BASED).calcPath(0, 3);